Prevented the deletion of evaluation that are in use (as in, events are created for these). Fix #168.

+++ app/controllers/mixevals_controller.php
@@ -271,6 +271,20 @@ class MixevalsController extends AppController
 
 	function delete($id)
 	{
+        // Deny Deleting evaluations in use:
+        $inUse = $this->Event->checkEvaluationToolInUse('4',$id);
+
+        if ($inUse) {
+            $this->index();
+            $message = "<span style='color:red'>";
+            $message.= "This evaluation is now in use, and can NOT be deleted.<br />";
+            $message.= "Please remove all the events assosiated with this evaluation first.";
+            $message.= "</span>";
+            $this->set('message', $message);
+            $this->render('index');
+            exit;
+        }
+
 		if ($this->Mixeval->del($id))
 		{
 			$this->MixevalsQuestion->deleteQuestions($id);
